Extract hook into own unit
